Time: 0 ms (100%), Space: 7.9 MB (33.82%) - LeetHub

+++ 0371-sum-of-two-integers/0371-sum-of-two-integers.cpp
@@ -1,9 +1,18 @@
 class Solution {
 public:
     int getSum(int a, int b) {
+        // ================================
+        // Approach 1: Bit Manipulation
+        // - Time  Complexity: O(1)
+        //   (Worst case 32 iterations.)
+        // - Space Complexity: O(1)
+        //   (Constant space used.)
+        // - Pros: Optimal.
+        // - Cons: Slightly tricky.
+        // ================================
         while (b) {
             int res = a ^ b;
-            int carry = (a & b) << 1;
+            unsigned int carry = (unsigned int)(a & b) << 1;
             a = res;
             b = carry;
         }
